semi_automation_data_loder/requirements.txt
boto3==1.24.28
pandas==1.4.4
Pillow==9.4.0
snowflake==0.0.3
snowflake_util==1.0.0b4
SQLAlchemy==1.4.39
streamlit==1.15.2
